redirections management remade, norm
